Refactor: Add BookmarkId Validation
bookmarkId 유효성 검사 추가. Null 또는 해당 bookmark가 없으면 Bad_Request 반환

by. chanho 2024-04-05

+++ src/main/java/com/movie/test/report/bookmark/controller/BookmarkController.java
@@ -68,6 +68,9 @@ public class BookmarkController {
     @DeleteMapping("/delete")
     public ResponseEntity deleteBookmark(Long bookmarkId){
 
+        if (!bookmarkService.validationBookmarkId(bookmarkId)) {
+            return new ResponseEntity("잘못된 시도입니다. 다시 시도해주세요.", HttpStatus.BAD_REQUEST);
+        }
         bookmarkService.deleteBookmark(bookmarkId);
 
         return new ResponseEntity(HttpStatus.OK);
